requirements.txt
jsonpickle==1.2
Flask==1.0.2
Flask-Login==0.4.1
Werkzeug==0.14.1
google-cloud-storage